requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
apscheduler
sqlalchemy
alembic
psycopg[binary]
celery
redis
tweepy
openai
sendgrid
twilio
jinja2
python-multipart
pytest
pytest-env
httpx
requests
python-dotenv
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-httpx
feedparser
